PlayerInput + Timer class
added player input method, created timer
