+++ game.py
@@ -1,21 +1,14 @@
-import time
+from board import Board
+from timer import Timer
 class Game:
     '''Game class that keeps track of score and timer'''
 
     def __init__(self, limit):
         ''' input limit is 0 for score-based games, nonzero for timed games '''
         self.score = 0
-        self.time_limit = limit
-        self.time_start = 0
-    
-    def start_time(self):
-        ''' marks current time and stores as self.time_start'''
-        self.time_start = time.time()
-    
-    def get_time(self):
-        ''' if there is a time limit, returns the amount of time left.
-            otherwise, returns amount of time passed since game start '''
-        if self.time_limit != 0:
-            return self.time_limit - (time.time() - self.time_start)
-        else:
-            return time.time() - self.time_start
+        self.timer = Timer(limit)
+        self.board = Board()
+        
+    def getPlayerInput(self):
+        print(self.board)
+        return input("Enter a word: ")
